Updated to the latest version of Commander and added alias support.
Updated to Commander 1.1.1; added alias support; tested with the CLI;
all unit tests pass.
