+++ Readme.md
@@ -1,4 +1,4 @@
-# Commander.js
+# commander-plus
 
   The complete solution for [node.js](http://nodejs.org) command-line interfaces, inspired by Ruby's [commander](https://github.com/visionmedia/commander).
 
@@ -42,7 +42,7 @@ console.log('  - %s cheese', program.cheese);
 
  The help information is auto-generated based on the information commander already knows about your program, so the following `--help` info is for free:
 
-```  
+```
  $ ./examples/pizza --help
 
    Usage: pizza [options]
@@ -227,6 +227,14 @@ program.choose(list, function(i){
   console.log('you chose %d "%s"', i, list[i]);
 });
 ```
+
+## .outputHelp()
+
+  Output help information without exiting.
+
+## .help()
+
+  Output help information and exit immediately.
 
 ## Links
