+++ examples/custom-help
@@ -29,4 +29,4 @@ program.on('--help', function(){
 
 program.parse(process.argv);
 
-console.log('stuff');
+program.help();
