+++ examples/pizza
@@ -17,7 +17,7 @@ program
 
 console.log('you ordered a pizza with:');
 if (program.peppers) console.log('  - peppers');
-if (program.pineappe) console.log('  - pineapple');
+if (program.pineapple) console.log('  - pineapple');
 if (program.bbq) console.log('  - bbq');
 
 var cheese = true === program.cheese
